Adapt rebalancing frequency with AIMD logic

+++ src/utils.cpp
@@ -1,6 +1,5 @@
 #include <thread>
-#include <volimem/x86constants.h>
 
 #include "utils.h"
 
-void sleep_ms(usize ms) { std::this_thread::sleep_for(Milliseconds(ms)); }
+void sleep_ms(u64 ms) { std::this_thread::sleep_for(Milliseconds(ms)); }
